Updating my drill branch

+++ exec/java-exec/src/test/java/org/apache/drill/exec/store/parquet/columnreaders/TestBatchSizingMemoryUtil.java
@@ -33,7 +33,6 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 public class TestBatchSizingMemoryUtil extends PhysicalOpUnitTestBase {
-//  private static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(TestBatchSizingMemoryUtil.class);
 
   // Batch schema
   private static TupleMetadata schema;
@@ -93,9 +92,9 @@ public class TestBatchSizingMemoryUtil extends PhysicalOpUnitTestBase {
 
     for (int columnIdx = 0; columnIdx < 3; columnIdx++) {
       final ColumnMemoryUsageInfo columnInfo = columnMemoryInfo[columnIdx];
-      final int remainingBitsCapacity = getRemainingBitsCapacity(columnInfo);
-      final int remainingOffsetsCapacity = getRemainingOffsetsCapacity(columnInfo);
-      final int remainingDataCapacity = getRemainingDataCapacity(columnInfo);
+      final long remainingBitsCapacity = getRemainingBitsCapacity(columnInfo);
+      final long remainingOffsetsCapacity = getRemainingOffsetsCapacity(columnInfo);
+      final long remainingDataCapacity = getRemainingDataCapacity(columnInfo);
 
       // Test current VV is within quota (since we are not adding new entries)
       Assert.assertTrue(BatchSizingMemoryUtil.canAddNewData(columnInfo, 0, 0, 0));
@@ -152,15 +151,15 @@ public class TestBatchSizingMemoryUtil extends PhysicalOpUnitTestBase {
     return result;
   }
 
-  private static int getRemainingBitsCapacity(ColumnMemoryUsageInfo columnInfo) {
+  private static long getRemainingBitsCapacity(ColumnMemoryUsageInfo columnInfo) {
     return columnInfo.vectorMemoryUsage.bitsBytesCapacity - columnInfo.vectorMemoryUsage.bitsBytesUsed;
   }
 
-  private static int getRemainingOffsetsCapacity(ColumnMemoryUsageInfo columnInfo) {
+  private static long getRemainingOffsetsCapacity(ColumnMemoryUsageInfo columnInfo) {
     return columnInfo.vectorMemoryUsage.offsetsByteCapacity - columnInfo.vectorMemoryUsage.offsetsBytesUsed;
   }
 
-  private static int getRemainingDataCapacity(ColumnMemoryUsageInfo columnInfo) {
+  private static long getRemainingDataCapacity(ColumnMemoryUsageInfo columnInfo) {
     return columnInfo.vectorMemoryUsage.dataByteCapacity - columnInfo.vectorMemoryUsage.dataBytesUsed;
   }
